+++ CHANGES.rst
@@ -19,6 +19,9 @@ Features
 
 - Add support for pytest 8.2, 8.3.
 
+- Add ``--fail-on-flaky`` option to fail the test run with custom exit code
+  when test passed on rerun.
+
 14.0 (2024-03-13)
 -----------------